[SERVER (sejin) / FEAT] 로비 수정
- FEAT: 로비 UI 추가 및 수정
- FIX: 로그인 - 로비 연동 시도
