+++ SERENDELE/Assets/Graphics/WizardInfo.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 70bd312f095c90746aa7847d69f514dd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
