+++ SERENDELE/Assets/Photon/PhotonUnityNetworking/Resources/PhotonServerSettings.asset
@@ -41,4 +41,4 @@ MonoBehaviour:
   - RespawnSpaceship
   DisableAutoOpenWizard: 1
   ShowSettings: 0
-  DevRegionSetOnce: 0
+  DevRegionSetOnce: 1
