--- SERENDELE/Assets/Prefabs/Weapon/Sphere.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 60b91e11f4a95b24ab1c95e2fa8f9658
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
