--- SERENDELE/Assets/Prefabs/Weapon/sword.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: fb60f3368443e4049b0ffadea4717263
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
