+++ SERENDELE/Assets/Resources/Wizard.prefab
@@ -1083,8 +1083,8 @@ GameObject:
   - component: {fileID: 1166922848213020171}
   - component: {fileID: 6330879752718203654}
   - component: {fileID: 2712990766098640499}
-  - component: {fileID: 5010154624989332907}
   - component: {fileID: 3648153412621398754}
+  - component: {fileID: 8737164284212867287}
   m_Layer: 0
   m_Name: Wizard
   m_TagString: Player
@@ -1237,27 +1237,11 @@ MonoBehaviour:
   OwnershipTransfer: 0
   observableSearch: 2
   ObservedComponents:
-  - {fileID: 5010154624989332907}
   - {fileID: 3648153412621398754}
+  - {fileID: 8737164284212867287}
   sceneViewId: 0
   InstantiationId: 0
   isRuntimeInstantiated: 0
---- !u!114 &5010154624989332907
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1480833682164876789}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 627855c7f81362d41938ffe0b1475957, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  m_SynchronizePosition: 1
-  m_SynchronizeRotation: 1
-  m_SynchronizeScale: 0
-  m_UseLocal: 1
 --- !u!114 &3648153412621398754
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1290,6 +1274,22 @@ MonoBehaviour:
     LayerIndex: 0
   - SynchronizeType: 0
     LayerIndex: 1
+--- !u!114 &8737164284212867287
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1480833682164876789}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 627855c7f81362d41938ffe0b1475957, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_SynchronizePosition: 1
+  m_SynchronizeRotation: 1
+  m_SynchronizeScale: 0
+  m_UseLocal: 1
 --- !u!1 &1536895868903762663
 GameObject:
   m_ObjectHideFlags: 0
