+++ SERENDELE/Assets/Scenes/SEJIN/2. CreateServer.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1009a20bb7da5f242b00231139ab25ab
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
