--- SERENDELE/Assets/Scenes/SEJIN/CreateServer.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 1009a20bb7da5f242b00231139ab25ab
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
