+++ SERENDELE/Assets/Scenes/SEJIN/_SEJIN.unity
@@ -421,7 +421,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 383811161}
   serializedVersion: 2
-  m_LocalRotation: {x: -0.0828865, y: -0.000000029699775, z: -0.0000000024702105, w: 0.996559}
+  m_LocalRotation: {x: -0.08288653, y: -0.000000012302052, z: -0.0000000010231951, w: 0.996559}
   m_LocalPosition: {x: 1.64, y: 1, z: 0.3}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
@@ -1065,7 +1065,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1060107930}
   serializedVersion: 2
-  m_LocalRotation: {x: -0.0828865, y: -0.000000029699775, z: -0.0000000024702105, w: 0.996559}
+  m_LocalRotation: {x: -0.08288653, y: -0.000000012302052, z: -0.0000000010231951, w: 0.996559}
   m_LocalPosition: {x: 1.64, y: 1, z: 0.3}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
@@ -1450,6 +1450,10 @@ PrefabInstance:
     - target: {fileID: 1480833682164876789, guid: bbd68a12556671e43862f167670a528f, type: 3}
       propertyPath: m_Name
       value: Wizard
+      objectReference: {fileID: 0}
+    - target: {fileID: 3278284138245518011, guid: bbd68a12556671e43862f167670a528f, type: 3}
+      propertyPath: m_Enabled
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 7090133079756509894, guid: bbd68a12556671e43862f167670a528f, type: 3}
       propertyPath: m_AnchorMax.x
