+++ SERENDELE/Assets/Scripts/Server/LobbyManager.cs
@@ -7,23 +7,49 @@ using UnityEngine.UI;
 
 public class LobbyManager : MonoBehaviourPunCallbacks
 {
+    [SerializeField] GameObject serverUI; // UI 캔버스
+
+    // 룸 생성 및 입장 관련 UI
     [SerializeField] TMP_Text nickName;
     [SerializeField] TMP_InputField input_RoomName;
-    [SerializeField] TMP_InputField input_MaxPlayer;
     [SerializeField] Button btn_CreateRoom;
     [SerializeField] Button btn_JoinRoom;
     [SerializeField] GameObject roomListItem;
+
+    // 캐릭터 선택 관련 UI
+    [SerializeField] Button btn_Arie;
+    [SerializeField] GameObject infoArie;
+
+    [SerializeField] Button btn_Lembra;
+    [SerializeField] GameObject infoLembra;
+
+    [SerializeField] bool Lembra;
+    [SerializeField] bool Arie;
+
+    // 경고창
+    [SerializeField] Image warnningMsg;
+    [SerializeField] TMP_Text chooseCha; // 캐릭터 선정하지 않고 버튼 누를 시
+    [SerializeField] TMP_Text inputTxt;  // 방 이름 적지 않고 Create 버튼 누를 시 
+    [SerializeField] TMP_Text clickList; // 방 선택하지 않고 Enter 버튼 누를 시
+
+
+
     public Transform rtContent;
 
     Dictionary<string, RoomInfo> dicRoomInfo = new Dictionary<string, RoomInfo>();
 
     void Start()
     {
-        nickName.text = PhotonNetwork.LocalPlayer.NickName;
+        //nickName.text = PhotonNetwork.LocalPlayer.NickName;
         input_RoomName.onValueChanged.AddListener(OnNameValueChanged);
-        input_MaxPlayer.onValueChanged.AddListener(OnPlayerValueChange);
         btn_CreateRoom.onClick.AddListener(OnClickCreateRoom);
         btn_JoinRoom.onClick.AddListener(OnClickJoinRoom);
+
+        infoArie.SetActive(false);
+        infoLembra.SetActive(false);
+        btn_Arie.onClick.AddListener(ClickArie);
+        btn_Lembra.onClick.AddListener(ClickLembra);
+
         PhotonNetwork.ConnectUsingSettings();
     }
 
@@ -78,7 +104,6 @@ public class LobbyManager : MonoBehaviourPunCallbacks
     void OnNameValueChanged(string s)
     {
         btn_JoinRoom.interactable = s.Length > 0;
-        btn_CreateRoom.interactable = input_RoomName.text.Length > 0 && input_MaxPlayer.text.Length > 0;
     }
 
     void OnPlayerValueChange(string s)
@@ -88,12 +113,7 @@ public class LobbyManager : MonoBehaviourPunCallbacks
 
     public void OnClickCreateRoom()
     {
-        int maxPlayers = int.Parse(input_MaxPlayer.text);
-        if (maxPlayers < 1 || maxPlayers > 5)
-        {
-            Debug.LogError("Max Players must be between 1 and 5");
-            return;
-        }
+        int maxPlayers = 2;
 
         RoomOptions options = new RoomOptions
         {
@@ -126,11 +146,39 @@ public class LobbyManager : MonoBehaviourPunCallbacks
     {
         base.OnJoinedRoom();
         Debug.Log("Joined room successfully");
+        GameObject player = PhotonNetwork.Instantiate("Wizard", Vector3.zero, Quaternion.identity);
+
+        if (serverUI != null)
+        {
+            serverUI.SetActive(false);
+        }
     }
 
     public override void OnJoinRoomFailed(short returnCode, string message)
     {
         base.OnJoinRoomFailed(returnCode, message);
         Debug.Log("Failed to join room: " + message);
+    }
+
+    public override void OnPlayerEnteredRoom(Player newPlayer)
+    {
+        Debug.Log($"{newPlayer.NickName} has entered the room.");
+
+    }
+
+    void ClickArie()
+    {
+        Arie = true;
+        Lembra = false;
+        infoArie.SetActive(true);
+        infoLembra.SetActive(false);
+    }
+
+    void ClickLembra()
+    {
+        Arie = false;
+        Lembra = true;
+        infoArie.SetActive(false);
+        infoLembra.SetActive(true);
     }
 }
