--- SERENDELE/Assets/Scripts/Weapon/Projectile.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d4fccbb6fe1459a48974af6da349f480
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
