+++ SERENDELE/Assets/Scripts/Weapon/Projectile2.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0d1bb51c6f1ccb0498c84971d7fcf393
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
